Refactored structure to separate models from enums, updated import paths.
