+++ backend/src/main/java/com/spring/restaurantmanagementsystem/model/Role.java
@@ -1,5 +1,6 @@
 package com.spring.restaurantmanagementsystem.model;
 
+import com.spring.restaurantmanagementsystem.enums.RoleEnum;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
