+++ backend/src/main/java/com/spring/restaurantmanagementsystem/repository/RoleRepository.java
@@ -1,7 +1,8 @@
 package com.spring.restaurantmanagementsystem.repository;
 
+import com.spring.restaurantmanagementsystem.enums.AllergenEnum;
 import com.spring.restaurantmanagementsystem.model.Role;
-import com.spring.restaurantmanagementsystem.model.RoleEnum;
+import com.spring.restaurantmanagementsystem.enums.RoleEnum;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,4 +11,5 @@ import java.util.Optional;
 @Repository
 public interface RoleRepository extends JpaRepository<Role, Long> {
     Optional<Role> findByName(RoleEnum name);
+    boolean existsByName(RoleEnum name);
 }
